store and read number of nodes and cores per node on remote file
Since each Workflow job is created once and its settings can not change the number of nodes and cores per node must be stored in a remote Workflow job configuration file. I called this file WorkflowJobInfo.txt, the number of nodes is written on the first line and the number of cores per node is written on the second line.